buffers: bring back a transmute
The compiler refuses to perform this cast as the alignment does not
match. bindgen generates this struct with a u16/4-byte alignment which
does not match the structures in which it is embedded.

I have not found a way to convince bindgen to generate the appropriate
alignment so this brings back the transmute in order not to block the
other removals of transmute in this patch series.